Data types limitation
 - `unsigned int` were too short (0 to 65 535)
 - changes to  `unsigned long` (0 to 4 294 967 295)

+++ problems/00002.c
@@ -3,14 +3,14 @@
 #define MAX 4000000
 
 int main(){
-    unsigned int a = 1;
-    unsigned int b = 2;
-    unsigned int fib = 0;
+    unsigned long a = 1;
+    unsigned long b = 2;
+    unsigned long fib = 0;
     // sum initialized because the first 3 values (1, 1, 2) are fixed
-    unsigned int sum = 2;
+    unsigned long sum = 2;
     unsigned char i = 4;
 
-    while(a+b <MAX){
+    while(a+b < MAX){
         fib = a + b;
         a = b;
         b = fib;
@@ -18,9 +18,9 @@ int main(){
         if(fib%2 == 0){
             sum += fib;
         }
-        printf("%d: value:%u / sum:%u\n",i, fib, sum);
+        printf("%d: value:%lu / sum:%lu\n",i, fib, sum);
         i++;
     }
-    printf("\nRESULT: %u\n", sum); 
+    printf("\nRESULT: %lu\n", sum); 
     return 0;
 }
